CORE-4688 Updated onDragDrop handler to not copy Arguments.
Updated the ArgListEditorDropTarget#onDragDrop handler in ArgumentListEditor to use the Argument set in the event data rather than a copy.
This prevents TreeSelection type Arguments from losing their selection items, since those are contained in a root SelectionItemGroup in the Argument, but copying the Argument can only de-serialize the root as a SelectionItem in the copy.